+++ SimulationIOFilters/Utility/EntriesHelper.h
@@ -8,5 +8,6 @@
 
 namespace EntriesHelper
 {
-void writeEntries(QTextStream& stream, std::function<QString(int32_t)> func, int32_t maxIndex, size_t maxEntriesPerLine, const QString& delimiter = " ", const QString& eol = "\n", size_t startingNumEntries = 0);
+void writeEntries(QTextStream& stream, std::function<QString(size_t)> func, size_t maxIndex, size_t maxEntriesPerLine, const QString& delimiter = " ", const QString& eol = "\n",
+                  size_t startingNumEntries = 0);
 }
